requirements.txt
fastapi==0.109.0
openai_whisper==20231117
pydantic==2.5.3
pytube==15.0.0
Requests==2.31.0
streamlit==1.30.0
